build blob variable by blob builder.

+++ Samples~/Variable/ComponentVariableData.cs
@@ -5,7 +5,7 @@ namespace EntitiesBT.Sample
     [GenerateAuthoringComponent]
     public struct ComponentVariableData : IComponentData
     {
-        public int InputValue;
-        public float OutputValue;
+        public float InputValue;
+        public int OutputValue;
     }
 }
